requirements.txt
numpy
pandas
python-pptx
